refactor(hooks): Port hooks to events
