+++ lib/AppInfo/Application.php
@@ -11,12 +11,14 @@ namespace OCA\Guests\AppInfo;
 use OCA\Files\Event\LoadAdditionalScriptsEvent;
 use OCA\Guests\Capabilities;
 use OCA\Guests\GroupBackend;
-use OCA\Guests\Hooks;
+use OCA\Guests\Listener\BeforeFileSystemSetupListener;
 use OCA\Guests\Listener\BeforeTemplateRenderedListener;
 use OCA\Guests\Listener\BeforeUserManagementRenderedListener;
 use OCA\Guests\Listener\LoadAdditionalScriptsListener;
 use OCA\Guests\Listener\ShareAutoAcceptListener;
+use OCA\Guests\Listener\ShareCreatedListener;
 use OCA\Guests\Listener\UserChangedListener;
+use OCA\Guests\Listener\UserLoggedInListener;
 use OCA\Guests\Notifications\Notifier;
 use OCA\Guests\RestrictionManager;
 use OCA\Guests\UserBackend;
@@ -26,15 +28,13 @@ use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\AppFramework\Http\Events\BeforeTemplateRenderedEvent;
-use OCP\EventDispatcher\IEventDispatcher;
 use OCP\IGroupManager;
 use OCP\IUserManager;
 use OCP\IUserSession;
-use OCP\Notification\IManager as INotificationManager;
 use OCP\Share\Events\ShareCreatedEvent;
 use OCP\User\Events\UserChangedEvent;
 use OCP\User\Events\UserFirstTimeLoggedInEvent;
-use Psr\Container\ContainerInterface;
+use OCP\User\Events\UserLoggedInEvent;
 
 class Application extends App implements IBootstrap {
 	public const APP_ID = 'guests';
@@ -51,6 +51,11 @@ class Application extends App implements IBootstrap {
 		$context->registerEventListener(BeforeTemplateRenderedEvent::class, BeforeTemplateRenderedListener::class);
 		$context->registerEventListener(BeforeUserManagementRenderedEvent::class, BeforeUserManagementRenderedListener::class);
 		$context->registerEventListener(UserChangedEvent::class, UserChangedListener::class);
+		$context->registerEventListener(UserLoggedInEvent::class, UserLoggedInListener::class);
+		$context->registerEventListener(ShareCreatedEvent::class, ShareCreatedListener::class);
+		$context->registerEventListener(UserFirstTimeLoggedInEvent::class, BeforeFileSystemSetupListener::class);
+
+		$context->registerNotifierService(Notifier::class);
 	}
 
 	public function boot(IBootContext $context): void {
@@ -59,39 +64,16 @@ class Application extends App implements IBootstrap {
 		$container->get(IUserManager::class)->registerBackend($container->get(UserBackend::class));
 		$container->get(IGroupManager::class)->addBackend($container->get(GroupBackend::class));
 
-		$this->setupGuestManagement($context->getAppContainer(), $context->getServerContainer());
-		$this->setupGuestRestrictions($context->getAppContainer(), $context->getServerContainer());
-		$this->setupNotifications($context->getAppContainer());
-		$context->getAppContainer()->get(RestrictionManager::class)->lateSetupRestrictions();
-	}
+		// Setup guest restrictions
+		$context->injectFn(function (IUserSession $userSession, RestrictionManager $restrictionManager): void {
+			$user = $userSession->getUser();
 
-	private function setupGuestManagement(ContainerInterface $container, ContainerInterface $server): void {
-		/** @var Hooks $hookManager */
-		$hookManager = $container->get(Hooks::class);
-		$server->get(IEventDispatcher::class)->addListener(ShareCreatedEvent::class, $hookManager->handlePostShare(...));
-		$server->get(IEventDispatcher::class)->addListener(UserFirstTimeLoggedInEvent::class, $hookManager->handleFirstLogin(...));
-	}
-
-	private function setupGuestRestrictions(ContainerInterface $container, ContainerInterface $server): void {
-		/** @var IUserSession $userSession */
-		$userSession = $server->get(IUserSession::class);
-		$user = $userSession->getUser();
-		/** @var RestrictionManager $restrictionManager */
-		$restrictionManager = $container->get(RestrictionManager::class);
-
-		if ($user) {
-			$restrictionManager->verifyAccess();
-			$restrictionManager->setupRestrictions();
-		} else {
-			$userSession->listen('\OC\User', 'postLogin', function () use ($restrictionManager): void {
+			if ($user) {
 				$restrictionManager->verifyAccess();
 				$restrictionManager->setupRestrictions();
-			});
-		}
-	}
+			}
 
-	private function setupNotifications(ContainerInterface $container): void {
-		$notificationManager = $container->get(INotificationManager::class);
-		$notificationManager->registerNotifierService(Notifier::class);
+			$restrictionManager->lateSetupRestrictions();
+		});
 	}
 }
