--- lib/Hooks.php
@@ -1,152 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * SPDX-FileCopyrightText: 2017-2024 Nextcloud GmbH and Nextcloud contributors
- * SPDX-FileCopyrightText: 2017 ownCloud GmbH
- * SPDX-License-Identifier: AGPL-3.0-only AND (AGPL-3.0-or-later OR AGPL-3.0-only)
- */
-
-namespace OCA\Guests;
-
-use OC\Files\Filesystem;
-use OCA\Guests\AppInfo\Application;
-use OCA\Guests\Service\InviteService;
-use OCA\Guests\Storage\ReadOnlyJail;
-use OCP\Constants;
-use OCP\Files\Storage\IStorage;
-use OCP\IConfig;
-use OCP\IUserManager;
-use OCP\IUserSession;
-use OCP\Share\Events\ShareCreatedEvent;
-use OCP\Share\IShare;
-use OCP\User\Events\UserFirstTimeLoggedInEvent;
-use Psr\Log\LoggerInterface;
-
-class Hooks {
-	public function __construct(
-		private readonly LoggerInterface $logger,
-		private readonly IUserSession $userSession,
-		private readonly IUserManager $userManager,
-		private readonly IConfig $config,
-		private readonly GuestManager $guestManager,
-		private readonly UserBackend $userBackend,
-		private readonly TransferService $transferService,
-		private readonly InviteService $inviteService,
-	) {
-	}
-
-	public function handlePostShare(ShareCreatedEvent $event): void {
-		$share = $event->getShare();
-
-		// Only process shares targeting a specific user or email (guest invite logic).
-		// Link shares, remote shares, etc. have no shareWith recipient and must be skipped.
-		$allowedShareTypes = [
-			IShare::TYPE_USER,
-			IShare::TYPE_EMAIL
-		];
-
-		if (!in_array($share->getShareType(), $allowedShareTypes, true)) {
-			return;
-		}
-
-		$shareWith = $share->getSharedWith();
-
-		$isGuest = $this->guestManager->isGuest($shareWith);
-		if (!$isGuest) {
-			$this->logger->debug(
-				"ignoring user '" . $shareWith . "', not a guest",
-				['app' => Application::APP_ID]
-			);
-			return;
-		}
-
-		if ($share->getNodeType() !== 'folder' && $share->getNodeType() !== 'file') {
-			$this->logger->debug(
-				'ignoring share for itemType ' . $share->getNodeType(),
-				['app' => Application::APP_ID]
-			);
-			return;
-		}
-
-		$user = $this->userSession->getUser();
-		$this->userManager->get($shareWith);
-
-		if (!$user) {
-			throw new \Exception(
-				'post_share hook triggered without user in session'
-			);
-		}
-
-		$this->logger->debug("checking if '" . $shareWith . "' has a password",
-			['app' => Application::APP_ID]);
-
-		$uid = $user->getUID();
-
-		$this->inviteService->sendInvite($uid, $shareWith, $share);
-	}
-
-	/**
-	 * @param array<string, mixed> $params
-	 */
-	public function setupReadonlyFilesystem(array $params): void {
-		$uid = $params['user'];
-		$user = $this->userManager->get($uid);
-
-		if ($user && $this->guestManager->isGuest($user)) {
-			Filesystem::addStorageWrapper('guests.readonly', function ($mountPoint, IStorage $storage) use ($uid): ReadOnlyJail|IStorage {
-				if ($mountPoint === sprintf('/%s/', $uid)) {
-					return new ReadOnlyJail([
-						'storage' => $storage,
-						'mask' => Constants::PERMISSION_READ,
-						'path' => 'files'
-					]);
-				}
-
-				return $storage;
-			});
-		}
-	}
-
-	public function handleFirstLogin(UserFirstTimeLoggedInEvent $event): void {
-		if (!$this->config->getSystemValueBool('migrate_guest_user_data', false)) {
-			return;
-		}
-
-		$user = $event->getUser();
-		$this->logger->debug('User ' . $user->getUID() . ' logged in for the very first time. Checking guests data import.');
-
-		$email = $user->getEMailAddress();
-		if ($email === null) {
-			$this->logger->info('User ' . $user->getUID() . ' does not have an email address set. Skipping guests data import.');
-			return;
-		}
-
-		if (!$this->userBackend->userExists($email)) {
-			$this->logger->info('No guest user for ' . $email . ' found. Skipping guests data import.');
-			return;
-		}
-
-		if (strtolower($email) === strtolower($user->getUID())) {
-			// This is the guest user, logging in for the very first time
-			return;
-		}
-
-		$guestUser = $this->userManager->get($email);
-		if ($guestUser === null) {
-			$this->logger->warning('Guest user ' . $email . ' does not exist (anymore)');
-			return;
-		}
-
-		$this->transferService->transfer($guestUser, $user);
-
-		if (!$this->config->getSystemValueBool('remove_guest_account_on_conversion', false)) {
-			// Disable previous account
-			$guestUser->setEnabled(false);
-		} else {
-			// Remove previous account
-			$guestUser->delete();
-		}
-	}
-}
