+++ lib/Listener/UserLoggedInListener.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * SPDX-FileCopyrightText: 2025 Nextcloud GmbH
+ * SPDX-FileContributor: Carl Schwan
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+namespace OCA\Guests\Listener;
+
+use OCA\Guests\RestrictionManager;
+use OCP\EventDispatcher\Event;
+use OCP\EventDispatcher\IEventListener;
+use OCP\User\Events\UserLoggedInEvent;
+
+/**
+ * @template-implements IEventListener<UserLoggedInEvent>
+ */
+class UserLoggedInListener implements IEventListener {
+	public function __construct(
+		private RestrictionManager $restrictionManager,
+	) {
+	}
+	public function handle(Event $event): void {
+		if (!$event instanceof UserLoggedInEvent) {
+			return;
+		}
+
+		$this->restrictionManager->verifyAccess();
+		$this->restrictionManager->setupRestrictions();
+	}
+}
