+++ lib/RestrictionManager.php
@@ -12,7 +12,10 @@ namespace OCA\Guests;
 use OC\AppConfig;
 use OC\NavigationManager;
 use OCA\Files_External\Config\ExternalMountPoint;
+use OCA\Guests\Listener\BeforeFileSystemSetupListener;
+use OCP\EventDispatcher\IEventDispatcher;
 use OCP\Files\Config\IMountProviderCollection;
+use OCP\Files\Events\BeforeFileSystemSetupEvent;
 use OCP\Files\Mount\IMountPoint;
 use OCP\INavigationManager;
 use OCP\IRequest;
@@ -21,7 +24,6 @@ use OCP\IUser;
 use OCP\IUserSession;
 use OCP\Server;
 use OCP\Settings\IManager;
-use OCP\Util;
 use Psr\Log\LoggerInterface;
 
 class RestrictionManager {
@@ -31,12 +33,12 @@ class RestrictionManager {
 		private readonly IRequest $request,
 		private readonly IUserSession $userSession,
 		private readonly IServerContainer $server,
-		private readonly Hooks $hooks,
 		private readonly GuestManager $guestManager,
 		private readonly IMountProviderCollection $mountProviderCollection,
 		private readonly Config $config,
 		private readonly UserBackend $userBackend,
 		private readonly LoggerInterface $logger,
+		private readonly IEventDispatcher $eventDispatcher,
 	) {
 	}
 
@@ -54,7 +56,8 @@ class RestrictionManager {
 		}
 
 		if ($this->guestManager->isGuest($user)) {
-			Util::connectHook('OC_Filesystem', 'preSetup', $this->hooks, 'setupReadonlyFilesystem');
+			$this->eventDispatcher->addServiceListener(BeforeFileSystemSetupEvent::class, BeforeFileSystemSetupListener::class);
+
 			if (!$this->config->allowExternalStorage()) {
 				$this->mountProviderCollection->registerMountFilter(fn (IMountPoint $mountPoint, IUser $user): bool => !($mountPoint instanceof ExternalMountPoint && $this->guestManager->isGuest($user)));
 			}
